SQUASH: refactoring utils
